refactor: 예약 대기 순번 조회의 N+1을 단일 쿼리로 개선
